+++ app/Http/Controllers/BankAccountController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Requests\BankAccountsRequest;
 use App\Models\BankAccount;
 use App\Models\BankType;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use PDF;
@@ -123,13 +124,17 @@ class BankAccountController extends Controller
      */
     public function index(Request $request)
     {
-        if ($request->ajax()) {
-            $row = $request["row"];
-            return self::index_fetch_bank_account($row, 200, '');
+        if (Gate::allows('bank_accounts')) {
+            if ($request->ajax()) {
+                $row = $request["row"];
+                return self::index_fetch_bank_account($row, 200, '');
+            }
+            $banks_types = BankType::orderBy('bank_name')->get();
+            return view('taarife-payeh/bank-accounts.index')
+                ->with('banks_types', $banks_types);
+        } else {
+            return abort(401);
         }
-        $banks_types = BankType::orderBy('bank_name')->get();
-        return view('taarife-payeh/bank-accounts.index')
-            ->with('banks_types', $banks_types);
     }
 
     /**
@@ -150,19 +155,23 @@ class BankAccountController extends Controller
      */
     public function store(BankAccountsRequest $request)
     {
-        $bank_accounts = new BankAccount();
-        $bank_accounts->chk_active = $request->chk_active;
-        $bank_accounts->account_type = $request->input('account_type');
-        $bank_accounts->account_number = $request->input('account_number');
-        $bank_accounts->shaba_number = $request->input('shaba_number');
-        $bank_accounts->cart_number = $request->input('cart_number');
-        $bank_accounts->branch_name = $request->input('branch_name');
-        $bank_accounts->branch_address = $request->input('branch_address');
-        $bank_accounts->cheque_print_type = $request->input('cheque_print_type');
-        $bank_accounts->bank_type()->associate($request->bank_type);
-        $bank_accounts->save();
-        $row = $request["row"];
-        return self::index_fetch_bank_account($row, 200, 'حساب بانکی جدید ذخیره شد');
+        if (Gate::allows('bank_accounts')) {
+            $bank_accounts = new BankAccount();
+            $bank_accounts->chk_active = $request->chk_active;
+            $bank_accounts->account_type = $request->input('account_type');
+            $bank_accounts->account_number = $request->input('account_number');
+            $bank_accounts->shaba_number = $request->input('shaba_number');
+            $bank_accounts->cart_number = $request->input('cart_number');
+            $bank_accounts->branch_name = $request->input('branch_name');
+            $bank_accounts->branch_address = $request->input('branch_address');
+            $bank_accounts->cheque_print_type = $request->input('cheque_print_type');
+            $bank_accounts->bank_type()->associate($request->bank_type);
+            $bank_accounts->save();
+            $row = $request["row"];
+            return self::index_fetch_bank_account($row, 200, 'حساب بانکی جدید ذخیره شد');
+        } else {
+            return abort(401);
+        }
     }
 
     /**
@@ -184,17 +193,21 @@ class BankAccountController extends Controller
      */
     public function edit($id)
     {
-        $bank_accounts = BankAccount::find($id);
-        if ($bank_accounts) {
-            return response()->json([
-                'status' => 200,
-                'bank_accounts' => $bank_accounts,
-            ]);
+        if (Gate::allows('bank_accounts')) {
+            $bank_accounts = BankAccount::find($id);
+            if ($bank_accounts) {
+                return response()->json([
+                    'status' => 200,
+                    'bank_accounts' => $bank_accounts,
+                ]);
+            } else {
+                return response()->json([
+                    'status' => 404,
+                    'message' => 'حساب بانکی یافت نشد',
+                ]);
+            }
         } else {
-            return response()->json([
-                'status' => 404,
-                'message' => 'حساب بانکی یافت نشد',
-            ]);
+            return abort(401);
         }
     }
 
@@ -207,26 +220,30 @@ class BankAccountController extends Controller
      */
     public function update(BankAccountsRequest $request, $id)
     {
-        $bank_accounts = BankAccount::find($id);
-        if ($bank_accounts) {
-            $bank_accounts->chk_active = $request->chk_active;
-            $bank_accounts->account_type = $request->input('account_type');
-            $bank_accounts->account_number = $request->input('account_number');
-            $bank_accounts->shaba_number = $request->input('shaba_number');
-            $bank_accounts->cart_number = $request->input('cart_number');
-            $bank_accounts->branch_name = $request->input('branch_name');
-            $bank_accounts->branch_address = $request->input('branch_address');
-            $bank_accounts->cheque_print_type = $request->input('cheque_print_type');
-            $bank_accounts->fill($request->only(['account_type', 'account_number', 'shaba_number', 'cart_number', 'bank_name', 'branch_name', 'branch_address', 'cheque_print_type'])); // 'cover nabayad dashte bashe choon dar virayesh ax moshkel be vojood miyad'
-            $bank_accounts->bank_type()->associate($request->bank_type);
-            $bank_accounts->save();
-            $row = $request["row"];
-            return self::index_fetch_bank_account($row, 200, 'حساب بانکی ویرایش شد');
+        if (Gate::allows('bank_accounts')) {
+            $bank_accounts = BankAccount::find($id);
+            if ($bank_accounts) {
+                $bank_accounts->chk_active = $request->chk_active;
+                $bank_accounts->account_type = $request->input('account_type');
+                $bank_accounts->account_number = $request->input('account_number');
+                $bank_accounts->shaba_number = $request->input('shaba_number');
+                $bank_accounts->cart_number = $request->input('cart_number');
+                $bank_accounts->branch_name = $request->input('branch_name');
+                $bank_accounts->branch_address = $request->input('branch_address');
+                $bank_accounts->cheque_print_type = $request->input('cheque_print_type');
+                $bank_accounts->fill($request->only(['account_type', 'account_number', 'shaba_number', 'cart_number', 'bank_name', 'branch_name', 'branch_address', 'cheque_print_type'])); // 'cover nabayad dashte bashe choon dar virayesh ax moshkel be vojood miyad'
+                $bank_accounts->bank_type()->associate($request->bank_type);
+                $bank_accounts->save();
+                $row = $request["row"];
+                return self::index_fetch_bank_account($row, 200, 'حساب بانکی ویرایش شد');
+            } else {
+                return response()->json([
+                    'status' => 404,
+                    'message' => 'اطلاعاتی یافت نشد',
+                ]);
+            }
         } else {
-            return response()->json([
-                'status' => 404,
-                'message' => 'اطلاعاتی یافت نشد',
-            ]);
+            return abort(401);
         }
     }
 
@@ -238,8 +255,12 @@ class BankAccountController extends Controller
      */
     public function destroy($id)
     {
-        $bank_accounts = BankAccount::find($id);
-        $bank_accounts->delete();
-        return self::index_fetch_bank_account(10, 200, 'حساب بانکی حذف شد');
+        if (Gate::allows('bank_accounts')) {
+            $bank_accounts = BankAccount::find($id);
+            $bank_accounts->delete();
+            return self::index_fetch_bank_account(10, 200, 'حساب بانکی حذف شد');
+        } else {
+            return abort(401);
+        }
     }
 }
