+++ app/Http/Controllers/BankToFundController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\BankToFundRequest;
 use App\Models\BankAccount;
 use App\Models\BankToFund;
 use App\Models\BankType;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use PDF;
@@ -167,15 +168,19 @@ class BankToFundController extends Controller
      */
     public function index(Request $request)
     {
-        if ($request->ajax()) {
-            $row = $request["row"];
-            return self::index_fetch_bank_to_fund($row, 200, '');
+        if (Gate::allows('bank_to_fund')) {
+            if ($request->ajax()) {
+                $row = $request["row"];
+                return self::index_fetch_bank_to_fund($row, 200, '');
+            }
+            $banks_types = BankType::all();
+            $bank_accounts = BankAccount::all();
+            return view('financial-management/bank-to-fund.index')
+                ->with('banks_types', $banks_types)
+                ->with('bank_accounts', $bank_accounts);
+        } else {
+            return abort(401);
         }
-        $banks_types = BankType::all();
-        $bank_accounts = BankAccount::all();
-        return view('financial-management/bank-to-fund.index')
-            ->with('banks_types', $banks_types)
-            ->with('bank_accounts', $bank_accounts);
     }
 
     /**
@@ -196,21 +201,25 @@ class BankToFundController extends Controller
      */
     public function store(BankToFundRequest $request)
     {
-        $bank_to_fund = new BankToFund();
-        $bank_to_fund->form_date = $request->input('form_date');
-        $bank_to_fund->form_number = $request->input('form_number');
-        $bank_to_fund->cash_amount = str_replace(",", "", $request->input('cash_amount'));
-        $bank_to_fund->considerations1 = $request->input('considerations1');
-        $bank_to_fund->date = $request->input('date');
-        $bank_to_fund->deposit_amount = str_replace(",", "", $request->input('deposit_amount'));
-        $bank_to_fund->wage = str_replace(",", "", $request->input('wage'));
-        $bank_to_fund->issue_tracking = $request->input('issue_tracking');
-        $bank_to_fund->considerations2 = $request->input('considerations2');
-        $bank_to_fund->bank_type()->associate($request->bank);
-        $bank_to_fund->bank_account()->associate($request->bank_account_details);
-        $bank_to_fund->save();
-        $row = $request["row"];
-        return self::index_fetch_bank_to_fund($row, 200, 'از بانک به صندوق جدید ذخیره شد');
+        if (Gate::allows('bank_to_fund')) {
+            $bank_to_fund = new BankToFund();
+            $bank_to_fund->form_date = $request->input('form_date');
+            $bank_to_fund->form_number = $request->input('form_number');
+            $bank_to_fund->cash_amount = str_replace(",", "", $request->input('cash_amount'));
+            $bank_to_fund->considerations1 = $request->input('considerations1');
+            $bank_to_fund->date = $request->input('date');
+            $bank_to_fund->deposit_amount = str_replace(",", "", $request->input('deposit_amount'));
+            $bank_to_fund->wage = str_replace(",", "", $request->input('wage'));
+            $bank_to_fund->issue_tracking = $request->input('issue_tracking');
+            $bank_to_fund->considerations2 = $request->input('considerations2');
+            $bank_to_fund->bank_type()->associate($request->bank);
+            $bank_to_fund->bank_account()->associate($request->bank_account_details);
+            $bank_to_fund->save();
+            $row = $request["row"];
+            return self::index_fetch_bank_to_fund($row, 200, 'از بانک به صندوق جدید ذخیره شد');
+        } else {
+            return abort(401);
+        }
     }
 
     /**
@@ -232,17 +241,21 @@ class BankToFundController extends Controller
      */
     public function edit($id)
     {
-        $bank_to_fund = BankToFund::find($id);
-        if ($bank_to_fund) {
-            return response()->json([
-                'status' => 200,
-                'bank_to_fund' => $bank_to_fund,
-            ]);
+        if (Gate::allows('bank_to_fund')) {
+            $bank_to_fund = BankToFund::find($id);
+            if ($bank_to_fund) {
+                return response()->json([
+                    'status' => 200,
+                    'bank_to_fund' => $bank_to_fund,
+                ]);
+            } else {
+                return response()->json([
+                    'status' => 404,
+                    'message' => 'از بانک به صندوق یافت نشد',
+                ]);
+            }
         } else {
-            return response()->json([
-                'status' => 404,
-                'message' => 'از بانک به صندوق یافت نشد',
-            ]);
+            return abort(401);
         }
     }
 
@@ -255,27 +268,31 @@ class BankToFundController extends Controller
      */
     public function update(BankToFundRequest $request, $id)
     {
-        $bank_to_fund = BankToFund::find($id);
-        if ($bank_to_fund) {
-            $bank_to_fund->form_date = $request->input('form_date');
-            $bank_to_fund->form_number = $request->input('form_number');
-            $bank_to_fund->cash_amount = str_replace(",", "", $request->input('cash_amount'));
-            $bank_to_fund->considerations1 = $request->input('considerations1');
-            $bank_to_fund->date = $request->input('date');
-            $bank_to_fund->deposit_amount = str_replace(",", "", $request->input('deposit_amount'));
-            $bank_to_fund->wage = str_replace(",", "", $request->input('wage'));
-            $bank_to_fund->issue_tracking = $request->input('issue_tracking');
-            $bank_to_fund->considerations2 = $request->input('considerations2');
-            $bank_to_fund->bank_type()->associate($request->bank);
-            $bank_to_fund->bank_account()->associate($request->bank_account_details);
-            $bank_to_fund->update();
-            $row = $request["row"];
-            return self::index_fetch_bank_to_fund($row, 200, 'از بانک به صندوق ویرایش شد');
+        if (Gate::allows('bank_to_fund')) {
+            $bank_to_fund = BankToFund::find($id);
+            if ($bank_to_fund) {
+                $bank_to_fund->form_date = $request->input('form_date');
+                $bank_to_fund->form_number = $request->input('form_number');
+                $bank_to_fund->cash_amount = str_replace(",", "", $request->input('cash_amount'));
+                $bank_to_fund->considerations1 = $request->input('considerations1');
+                $bank_to_fund->date = $request->input('date');
+                $bank_to_fund->deposit_amount = str_replace(",", "", $request->input('deposit_amount'));
+                $bank_to_fund->wage = str_replace(",", "", $request->input('wage'));
+                $bank_to_fund->issue_tracking = $request->input('issue_tracking');
+                $bank_to_fund->considerations2 = $request->input('considerations2');
+                $bank_to_fund->bank_type()->associate($request->bank);
+                $bank_to_fund->bank_account()->associate($request->bank_account_details);
+                $bank_to_fund->update();
+                $row = $request["row"];
+                return self::index_fetch_bank_to_fund($row, 200, 'از بانک به صندوق ویرایش شد');
+            } else {
+                return response()->json([
+                    'status' => 404,
+                    'message' => 'اطلاعاتی یافت نشد',
+                ]);
+            }
         } else {
-            return response()->json([
-                'status' => 404,
-                'message' => 'اطلاعاتی یافت نشد',
-            ]);
+            return abort(401);
         }
     }
 
@@ -287,8 +304,12 @@ class BankToFundController extends Controller
      */
     public function destroy($id)
     {
-        $bank_to_fund = BankToFund::find($id);
-        $bank_to_fund->delete();
-        return self::index_fetch_bank_to_fund(10, 200, 'از بانک به صندوق حذف شد');
+        if (Gate::allows('bank_to_fund')) {
+            $bank_to_fund = BankToFund::find($id);
+            $bank_to_fund->delete();
+            return self::index_fetch_bank_to_fund(10, 200, 'از بانک به صندوق حذف شد');
+        } else {
+            return abort(401);
+        }
     }
 }
